ipc: float to int

+++ app/thread_gateway/src/protobuf/proto_deserializer.c
@@ -12,7 +12,7 @@ void deserialize_sensor_data(uint8_t *msg, uint8_t size, char *buffer)
         LOG_ERR("Decoding failed: %s\n", PB_GET_ERROR(&stream));
         return;
     }
-    sprintf(buffer, "ID %s, temp %d, humi %d ", log_strdup(message.sensor_id), (int)message.temperature, (int)message.humidity);
+    sprintf(buffer, "ID %s, temp %d, humi %d ", log_strdup(message.sensor_id), message.temperature, message.humidity);
 }
 
 void protobuf2str(uint8_t *data, size_t size, char *buffer)
